Removed DefaultReferenceResolver state from JsonSerializerInternalBase
